reduce hover highlight color, break out thumbnail code
